5.1.16 logged new operation request and 5.1.18 added validation where doctors only can delete requests they created

+++ Domain/OperationRequestAggr/OperationRequestService.cs
@@ -3,14 +3,15 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Sem5Pi2425.Domain.AppointmentAggr;
+using Sem5Pi2425.Domain.LogAggr;
 using Sem5Pi2425.Domain.OperationTypeAggr;
 using Sem5Pi2425.Domain.PatientAggr;
 using Sem5Pi2425.Domain.Shared;
 using Sem5Pi2425.Domain.StaffAggr;
 using Sem5Pi2425.Domain.SystemUserAggr;
+using Type = Sem5Pi2425.Domain.LogAggr.Type;
 
-namespace Sem5Pi2425.Domain.OperationRequestAggr
-{
+namespace Sem5Pi2425.Domain.OperationRequestAggr {
 
     public class OperationRequestService
     {
@@ -21,10 +22,10 @@ namespace Sem5Pi2425.Domain.OperationRequestAggr
         private readonly IPatientRepository _patientRepo;
         private readonly IUserRepository _userRepo;
         private readonly IStaffRepository _staffRepo;
+        private readonly LogService _logService;
 
-        public OperationRequestService(IUnitOfWork unitOfWork, IOperationRequestRepository repo, IAppointmentRepository appointmentRepo, 
-            IOperationTypeRepository operationTypeRepo, IPatientRepository patientRepo, IUserRepository userRepo, IStaffRepository staffRepository)
-        {
+        public OperationRequestService(IUnitOfWork unitOfWork, IOperationRequestRepository repo, IAppointmentRepository appointmentRepo, IOperationTypeRepository operationTypeRepo, 
+            IPatientRepository patientRepo, IUserRepository userRepo, IStaffRepository staffRepository, LogService logService) {
             _unitOfWork = unitOfWork;
             _repo = repo;
             _appointmentRepo = appointmentRepo;
@@ -32,8 +33,8 @@ namespace Sem5Pi2425.Domain.OperationRequestAggr
             _patientRepo = patientRepo;
             _userRepo = userRepo;
             _staffRepo = staffRepository;
+            _logService = logService;
         }
-
         
         public async Task<List<OperationRequestDTO>> GetAllOperationRequestsAsync() {
             var list = await _repo.GetAllAsync();
@@ -56,9 +57,9 @@ namespace Sem5Pi2425.Domain.OperationRequestAggr
         
         public async Task<ActionResult<OperationRequestDTO>> AddOperationRequestAsync(CreateOperationRequestDto dto, string email)
         {
+            var log = "";
             var deadlineString = dto.Deadline;
             DateOnly deadline = DateOnly.Parse(deadlineString);
-
             Enum.TryParse(dto.Priority, true, out Priority priority);
             
             var doctor = await _userRepo.GetByEmailAsync(email);
@@ -68,44 +69,53 @@ namespace Sem5Pi2425.Domain.OperationRequestAggr
             var operationType = await _operationTypeRepo.GetByIdAsync(dto.OperationType.Id);
             
             if (doctor == null) {
-                throw new Exception("Doctor not found");
+                throw new BusinessRuleValidationException("Doctor not found");
             }
             if (patient == null) {
-                throw new Exception("Patient not found");
+                throw new BusinessRuleValidationException("Patient not found");
             }
             if (operationType == null) {
-                throw new Exception("Operation type not found");
+                throw new BusinessRuleValidationException("Operation type not found");
             }
             
-            if (operationType.NeededSpecializations.Contains(doctorStaff.Specialization.ToString())) {
-                Console.WriteLine("OPERATION MATCHES");
+            if (operationType.NeededSpecializations.Contains(doctorStaff.Specialization.ToString()) && operationType.Active) {
                 var operationRequest = new OperationRequest(deadline,priority, doctor, patient, operationType);
 
-                await _repo.AddAsync(operationRequest);
-                await _unitOfWork.CommitAsync();
-
-                return new OperationRequestDTO(operationRequest.Id, operationRequest.Deadline, operationRequest.Priority,
-                    operationRequest.Doctor, operationRequest.Patient, operationRequest.OperationType);
-            }
+                    await _repo.AddAsync(operationRequest);
+                    await _unitOfWork.CommitAsync();
+                    log += $"NEW REQUEST ID: {operationRequest.Id.Value}; Deadline: {operationRequest.Deadline.ToString()}; Priority: {operationRequest.Priority}; " +
+                           $"Doctor: {operationRequest.Doctor.Username.Value}; Patient: {operationRequest.Patient.User.Username.Value}; Operation type: {operationRequest.OperationType.Name.Value}"; 
+                    _logService.AddLogAsync(new LogDto(null, Type.OperationRequest.ToString(), log, patient.Id.Value, null));
+                    return new OperationRequestDTO(operationRequest.Id, operationRequest.Deadline, operationRequest.Priority,
+                        operationRequest.Doctor, operationRequest.Patient, operationRequest.OperationType);
+            } 
+            
             throw new Exception("Operation type doesn't match the doctor’s specialization.");
+            
         }
         
-        public async Task<ActionResult<OperationRequestDTO>> DeleteOperationRequestAsync(OperationRequestId operationRequestId) {
+        public async Task<ActionResult<OperationRequestDTO>> DeleteOperationRequestAsync(OperationRequestId operationRequestId, string email) {
             var operationRequest = await _repo.GetByIdAsync(operationRequestId);
             var appointments = await _appointmentRepo.GetAllAsync();
+            var doctor = await _userRepo.GetByEmailAsync(email);
+            var operationDoctorId = operationRequest.Doctor.Id.Value;
 
-            foreach (var appointment in appointments)
+            if (operationDoctorId.Equals(doctor.Id.Value))
             {
-                if (appointment.OperationRequest.Equals(operationRequest))
+                foreach (var appointment in appointments)
                 {
-                    throw new Exception("Operation request already scheduled");
-                }
-            } 
-            _repo.Remove(operationRequest); 
-            await _unitOfWork.CommitAsync();
+                    if (appointment.OperationRequest.Equals(operationRequest))
+                    {
+                        throw new Exception("Operation request already scheduled");
+                    } 
+                } 
+                _repo.Remove(operationRequest); 
+                await _unitOfWork.CommitAsync();
             
-            return new OperationRequestDTO(operationRequest.Id, operationRequest.Deadline, operationRequest.Priority, 
-                operationRequest.Doctor, operationRequest.Patient, operationRequest.OperationType);
+                return new OperationRequestDTO(operationRequest.Id, operationRequest.Deadline, operationRequest.Priority, 
+                    operationRequest.Doctor, operationRequest.Patient, operationRequest.OperationType);
+            }
+            throw new Exception("Doctors can only delete operations requests they created");
         }
 
     }
